requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
google-generativeai
pypdf
python-dotenv
weaviate-client
pydantic

